chore: move system integration methods to a file
